requirements.txt
streamlit
phidata
pathlib
tempfile
google-generativeai
duckduckgosearch
dotenv
deep_translator
